+++ packages/schemas/package.json
@@ -6,7 +6,7 @@
   "main": "src/index.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node --enable-source-maps --no-warnings=ExperimentalWarning --loader ts-node/esm src/generators/index.ts"
+    "start": "node --enable-source-maps --no-warnings=ExperimentalWarning --loader ts-node/esm src/index.ts"
   },
   "dependencies": {
     "ajv": "^8.12.0",
